+++ src/briefy/leica/models/billing_info/customer.py
@@ -83,8 +83,13 @@ class CustomerBillingInfo(BillingInfo):
     Internal codes used by Finance to determine tax rates to be applied to this customer.
     """
 
-    def to_dict(self):
-        """Return a dict representation of this object."""
-        data = super().to_dict()
+    def to_dict(self, excludes: list=None, includes: list=None) -> dict:
+        """Return a dictionary with fields and values used by this Class.
+
+        :param excludes: attributes to exclude from dict representation.
+        :param includes: attributes to include from dict representation.
+        :returns: Dictionary with fields and values used by this Class
+        """
+        data = super().to_dict(excludes, includes)
         data['customer'] = self.customer.to_summary_dict()
         return data